Quick with timeseries (#46)
* созвон

* 2024-05-18 call

* refactor: Раскидал библиотечные вещи в основной код

* - warnings

* synthetic test fix

* comments

* Update README.md

* Update README.md 2

* Update README.md

* Update README.md

* Update README.md

---------
